Add support for individual file building

+++ src/mapdir.js
@@ -2,12 +2,30 @@ const {writeFileSync, mkdirSync, readdirSync, statSync, linkSync, renameSync, ex
 const {join, dirname} = require("path")
 const {sync: rimraf} = require("rimraf")
 
-exports.mapDir = function mapDir(source, dest, map, add = {}) {
+let nothing = Promise.resolve(undefined)
+
+exports.mapFile = function mapFile(name, fullPath, dest, map) {
+  let mapped = map(fullPath, name)
+  if (mapped && mapped.then) {
+    return mapped.then(result => {
+      writeFileSync(join(dest, result.name || name), result.content)
+    })
+  } else if (mapped) {
+    writeFileSync(join(dest, mapped.name || name), mapped.content)
+    return nothing
+  } else if (mapped !== false) {
+    linkSync(fullPath, join(dest, name))
+    return nothing
+  }
+}
+
+exports.mapDir = async function mapDir(source, dest, map, add = {}) {
   let above = join(dest, ".."), id = Math.floor(Math.random() * 0xffff).toString(16)
   let temp = join(above, "tmp-output-" + id)
   let waitFor = []
 
   function walkDir(dir, prefix) {
+    if (/tmp-output/.test(dir)) throw new Error("STOH")
     mkdirSync(join(temp, prefix), {recursive: true})
     
     for (let filename of readdirSync(dir)) {
@@ -17,40 +35,28 @@ exports.mapDir = function mapDir(source, dest, map, add = {}) {
       if (stat.isDirectory()) {
         walkDir(path, prefixed)
       } else {
-        let mapped = map(path, prefixed)
-        if (mapped && mapped.then) {
-          waitFor.push(mapped.then(result => {
-            writeFileSync(join(temp, result.name || prefixed), result.content)
-          }))
-        } else if (mapped) {
-          writeFileSync(join(temp, mapped.name || prefixed), mapped.content)
-        } else if (mapped !== false) {
-          linkSync(path, join(temp, prefixed))
-        }
+        let promise = exports.mapFile(prefixed, path, temp, map)
+        if (promise != nothing) waitFor.push(promise)
       }
     }
   }
 
+  let done = false
   try {
     walkDir(source, "")
     for (let added in add) {
       writeFileSync(join(temp, added), add[added])
     }
-  } catch(e) {
-    rimraf(temp)
-    throw e
-  }
-
-  Promise.all(waitFor).then(() => {
+    for (let p of waitFor) await p
     let moved
     if (existsSync(dest)) {
       moved = join(above, "moved-" + id)
       renameSync(dest, moved)
     }
     renameSync(temp, dest)
+    done = true
     if (moved) rimraf(moved)
-  }).catch(e => {
-    rimraf(temp)
-    throw e
-  })
+  } finally {
+    if (!done) rimraf(temp)
+  }
 }
